Made share UUID only work for get endpoint.
